backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0
PyJWT>=2.8.0
sqlalchemy>=2.0.0
chromadb>=0.5.0
openai>=1.10.0
pymupdf>=1.23.0
numpy>=1.26.0
aiofiles>=23.2.0
